requirements.txt
Flask_SQLAlchemy==2.5.1
Flask==2.0.1
PyJWT==2.3.0
